Added match tests from @howiefh/ant-path-matcher
