+++ .eslintrc
@@ -1,4 +1,11 @@
 {
     "root": true,
+    "parserOptions": {
+        "ecmaVersion": 2018,
+        "sourceType": "module"
+    },
+    "env": {
+        "es6": true
+    },
     "ignorePatterns": ["node_modules", ".*", "dist"]
 }
