+++ package.json
@@ -1,10 +1,11 @@
 {
   "name": "@maxbilbow/ant-path-matcher",
   "version": "0.0.0-rc.3",
+  "type": "module",
   "description": "An implementation for Ant-style path patterns",
-  "main": "dist/cjs/main.cjs",
-  "module": "dist/esm/main.js",
-  "types": "dist/types/main.d.ts",
+  "main": "./dist/cjs/main.cjs",
+  "module": "./dist/esm/main.js",
+  "types": "./dist/types/main.d.ts",
   "exports": {
     ".": {
       "types": "./dist/types/main.d.ts",
@@ -18,8 +19,9 @@
     "src"
   ],
   "scripts": {
+    "preversion": "npm run lint && npm test && npm run build",
     "lint": "eslint \"**/*.{ts,js,cjs}\"",
-    "test": "echo \"Error: no test specified\" && exit 1",
+    "test": "node --experimental-vm-modules ./node_modules/.bin/jest",
     "prepare": "husky",
     "prepublishOnly": "npm run build",
     "build": "tsc --build src/tsconfig.json && npm run build-esm && npm run build-cjs",
@@ -39,19 +41,21 @@
     "url": "https://github.com/cimc-raffles/ant-path-matcher.git"
   },
   "dependencies": {
-    "@tsconfig/node-lts": "^20.1.1",
-    "string-builder": "^0.1.8",
-    "xregexp": "^5.1.1",
-    "xregexp-quotemeta": "^1.0.7"
+    "xregexp": "^5.1.1"
   },
   "devDependencies": {
+    "@tsconfig/node-lts": "^20.1.1",
+    "@types/jest": "^29.5.12",
     "@typescript-eslint/eslint-plugin": "^6.20.0",
     "@typescript-eslint/parser": "^6.20.0",
     "esbuild": "^0.20.0",
     "eslint": "^8.56.0",
     "husky": "^9.0.10",
+    "jest": "^29.7.0",
     "lint-staged": "^15.2.1",
     "prettier": "^3.2.4",
+    "ts-jest": "^29.1.2",
+    "tsx": "^4.7.0",
     "typescript": "^5.3.3"
   },
   "lint-staged": {
